+++ doc/changelog.rst
@@ -37,6 +37,7 @@ Localization
 Admin
 ~~~~~
 
+- Add order address validation to admin order creator
 - Fix bug when editing anonymous orders
 - Show order line discount percentage in order detail and creator views
 - Allow superadmins to login as customer